+++ GUI.NET/Debugger/frmDebugger.cs
@@ -23,7 +23,7 @@ namespace Mesen.GUI.Debugger
 		private bool _firstBreak = true;
 		private bool _wasPaused = false;
 		private bool _executionIsStopped = false; //Flag used to break on the first instruction after power cycle/reset if execution was stopped before the reset
-		private int _previousCycle = 0;
+		private UInt64 _previousCycle = 0;
 
 		private InteropEmu.NotificationListener _notifListener;
 		private ICodeViewer _lastCodeWindow;
